backend/requirements.txt
fastapi[standard]==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
celery[redis]==5.4.0
redis==5.2.1
docling==2.26.0
langchain-docling
langchain-anthropic
langchain-voyageai
langchain-core
langchain-text-splitters
zvec
deepgram-sdk==3.9.0
httpx==0.28.1
aiofiles==24.1.0
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
